[IMP] account: Extract payment methods into their own file
The account payment file being quite big already,
and the payment methods being modified in another
task (2414749) we'll extract them into their own
file to make it easier to work with them

closes odoo/odoo#70937

Related: odoo/enterprise#18367
