+++ addons/account/models/__init__.py
@@ -16,6 +16,7 @@ from . import account_analytic_default
 from . import account_partial_reconcile
 from . import account_full_reconcile
 from . import account_payment
+from . import account_payment_method
 from . import account_bank_statement
 from . import chart_template
 from . import account_analytic_line
